Lesson changes
Added a function to write a table row of weekday abbreviations with a for loop to look through the table row array. Also added a function to calculate the number of days in the month by creating an array of the number of days in each month and extracting the four digit year and month value.